backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pandas
openpyxl
python-dotenv
google-generativeai
psycopg2-binary
python-multipart
xlrd